+++ api/src/routes/routes.ts
@@ -12,7 +12,7 @@ router
 	 * /:
 	 *   get:
 	 *     summary: API root
-	 *     description: Returns a welcome message for the bbai API and provides a link to the API documentation
+	 *     description: Returns a welcome message for the BBai API and provides a link to the API documentation
 	 *     responses:
 	 *       200:
 	 *         description: Successful response with welcome message and documentation link