+++ src/shared/config/configSchema.ts
@@ -32,6 +32,8 @@ export interface GlobalConfigSchema {
 	project: ProjectDataConfigSchema;
 	repoInfo: RepoInfoConfigSchema;
 	version: string;
+	bbaiExeName: string;
+	bbaiApiExeName: string;
 }
 
 export interface ProjectConfigSchema {
@@ -69,6 +71,8 @@ export const defaultGlobalConfig: GlobalConfigSchema = {
 		type: 'local',
 	},
 	version: 'unknown', // This will be overwritten by the actual version from version.ts
+	bbaiExeName: 'bbai', // This will be overwritten by the correct exe name when creating globalConfig
+	bbaiApiExeName: 'bbai-api', // This will be overwritten by the correct exe name when creating globalConfig
 };
 
 export const defaultProjectConfig: ProjectConfigSchema = {